+++ Sources/StreamVideo/Utils/Swift6Migration/CustomStringConvertible+Retroactive.swift
@@ -22,6 +22,7 @@ extension RTCSignalingState: @retroactive CustomStringConvertible {}
 extension RTCDataChannelState: @retroactive CustomStringConvertible {}
 extension RTCBundlePolicy: @retroactive CustomStringConvertible {}
 extension RTCContinualGatheringPolicy: @retroactive CustomStringConvertible {}
+extension CMSampleBuffer: @retroactive CustomStringConvertible {}
 #else
 extension AVAudioSession.Category: CustomStringConvertible {}
 extension AVAudioSession.CategoryOptions: CustomStringConvertible {}
@@ -38,4 +39,5 @@ extension RTCSignalingState: CustomStringConvertible {}
 extension RTCDataChannelState: CustomStringConvertible {}
 extension RTCBundlePolicy: CustomStringConvertible {}
 extension RTCContinualGatheringPolicy: CustomStringConvertible {}
+extension CMSampleBuffer: CustomStringConvertible {}
 #endif