+++ StreamVideoTests/Utils/Store/Store_PerformanceTests.swift
@@ -232,8 +232,8 @@ final class Store_PerformanceTests: XCTestCase, @unchecked Sendable {
         let iterations = 10000
 
         measure(
-            baseline: .init(local: 13, ci: 24, stringTransformer: { String(format: "%.4fs", $0) }),
-            allowedRegression: .init(local: 0.25, ci: 0.35),
+            baseline: .init(24, stringTransformer: { String(format: "%.4fs", $0) }),
+            allowedRegression: .init(local: 0.1, ci: 0.35),
             iterations: 2
         ) {
             // Wait for completion